Starts work on scene rendering

+++ Basic2D/Include/EOMSolver.h
@@ -8,9 +8,9 @@ class Vector;
 class EOMSolver {
 
 public:
-	static void resolve_time_step(World &world, unsigned int time_step);
-	static void simulate_object(Object &object, unsigned int time_step);
-	static Vector evaluate_forces(Object &object, unsigned int time_step);
+	static void resolve_time_step(World &world, double time_step);
+	static void simulate_object(Object &object, double time_step);
+	static Vector evaluate_forces(Object &object, double time_step);
 };
 
 #endif
